Allow sending values to the scaled fader before loadbang
Fixes a bug where, when sending a value to the scaled fader before it
has fully initialized, it would be lost. This was caused by the loadbang
being sent to the setup logic and the value cache at the same time, and
this fixes that by using a trigger to sequence the bangs (sets up the
expressions before attempting to use them).